#58 added file for GATTAnalogOutput
